Approximate source root for target
Previously, Fastpass would always use the target base directory as
source root. Some targets, however, have a parent directory as the
source root. This commit ports the logic that detects the source root
from Fastpass for Pants to Fastpass for Bazel.

This should fix issues where IntelliJ reports that the package name
doesn't correspond to the directory structure.
